Rename back to Permissions, but expose it through `liveblocks.createPermissions()` instead of a global import
